edits 1
You might want to add a note at the end of the introduction that this tutorial will work with the first Salt tutorial (and which step to skip to if you've already done the first one), but that it's not required since we will be starting from scratch in this tutorial. You might also want to note that following the first tutorial is not required when you link to it in the "Here's the plan" section.