edits 3
When you format commands that the user is going to run, can you make sure that they:
Have just the command, without a "#" in front of them
Are on separate lines for the most part, with a brief one-sentence introduction to what each one does. It's okay to combine commands like an update/upgrade or cd/download.
Please separate output from commands that the reader should enter

When you direct the reader to edit a configuration file, it's good to include the line as it should look in the final file. For example, this:

# vi /etc/salt/minion
# # set the appropriate line to read 'master: <^><your configmaster IP><^>'
# service salt-minion restart

Could become:

Open the `/etc/salt/minion` configuration file for editing:

vi /etc/salt/minion

Remove the comment from the `master` line, and add your master's IP address:

master: <^><your configmaster IP><^>

etc. Thanks!